[eRCP] Remove obsolete ColorSupport class
Primarily used to create Colors from the RGB values. The factory methods
for creating the ColorInfo objects has been moved to the
ColorPropertyEditor.

Contributes to
https://github.com/eclipse-windowbuilder/windowbuilder/issues/942
