+++ org.eclipse.wb.tests/src/org/eclipse/wb/tests/designer/swt/support/SupportTests.java
@@ -22,7 +22,6 @@ import org.junit.runners.Suite.SuiteClasses;
 @RunWith(Suite.class)
 @SuiteClasses({
 		CoordinateUtilsTest.class,
-		ColorSupportTest.class,
 		RowLayoutSupportTest.class,
 		ControlSupportTest.class,
 		ContainerSupportTest.class
